requirements.txt
langchain_huggingface==0.1.2
langchain==0.3.11
langchain_groq==0.2.1
langchain_core==0.3.28
langchain_community==0.3.11
faiss-cpu==1.9.0.post1
streamlit==1.35.0
pandas==2.1.4
python-dotenv==1.0.1
qrcode